change groupId to com.shopify, package to com.shopify.checkoutkit
