+++ lib/src/main/java/com/shopify/checkoutkit/CheckoutDialog.kt
@@ -20,7 +20,7 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.shopify.checkout
+package com.shopify.checkoutkit
 
 import android.app.Dialog
 import android.content.Context
