+++ lib/src/main/res/layout/dialog_checkout.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.AppCompat.DayNight"
         android:elevation="4dp" />
 
-    <com.shopify.checkout.CheckoutWebViewContainer
+    <com.shopify.checkoutkit.CheckoutWebViewContainer
         android:id="@+id/checkoutSdkContainer"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
