+++ lib/src/test/java/com/shopify/checkoutkit/CheckoutBridgeTest.kt
@@ -20,10 +20,10 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.shopify.checkout
+package com.shopify.checkoutkit
 
-import com.shopify.checkout.CheckoutBridge.CheckoutWebOperation.COMPLETED
-import com.shopify.checkout.CheckoutBridge.CheckoutWebOperation.MODAL
+import com.shopify.checkoutkit.CheckoutBridge.CheckoutWebOperation.COMPLETED
+import com.shopify.checkoutkit.CheckoutBridge.CheckoutWebOperation.MODAL
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import org.junit.Before
