+++ lib/src/test/java/com/shopify/checkoutkit/CheckoutDialogTest.kt
@@ -20,7 +20,7 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.shopify.checkout
+package com.shopify.checkoutkit
 
 import android.app.Dialog
 import android.graphics.drawable.ColorDrawable
