+++ samples/MobileBuyIntegration/app/src/main/java/com/shopify/checkout_sdk_mobile_buy_integration_sample/MobileBuyIntegration.kt
@@ -23,7 +23,7 @@
 package com.shopify.checkout_sdk_mobile_buy_integration_sample
 
 import android.app.Application
-import com.shopify.checkout.ShopifyCheckoutKit
+import com.shopify.checkoutkit.ShopifyCheckoutKit
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.di.setupDI
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.settings.PreferencesManager
 import kotlinx.coroutines.CoroutineScope
